Move files array processing into ServerRequestFactory.
Refs #14213
